+++ DL/Location.cs
@@ -11,9 +11,9 @@ public partial class Location
 
     public string? PlaceId { get; set; }
 
-    public double Latitude { get; set; }
+    public decimal Latitude { get; set; }
 
-    public double Longitude { get; set; }
+    public decimal Longitude { get; set; }
 
     public virtual ICollection<Order> OrderIdDropOffNavigations { get; set; } = new List<Order>();
